requirements.txt
amqp==5.0.6
asgiref==3.4.1
attrs==21.2.0
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
charset-normalizer==2.0.3
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.5
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-cors-headers==3.7.0
django-debug-toolbar==3.2.1
django-filter==2.4.0
django-json-widget==1.1.1
django-property-filter==1.1.0
django-rest-swagger==2.2.0
django-timezone-field==4.2.1
djangorestframework==3.12.4
drf-spectacular==0.17.2
et-xmlfile==1.1.0
Faker==8.12.0
future==0.18.2
idna==3.2
inflection==0.5.1
install==1.3.4
itypes==1.2.0
jdatetime==3.6.2
Jinja2==3.0.1
jsonschema==3.2.0
kombu==5.1.0
Markdown==3.3.4
MarkupSafe==2.0.1
numpy==1.21.2
openapi-codec==1.3.2
openpyxl==3.0.7
pandas==1.3.3
prompt-toolkit==3.0.19
pyrsistent==0.18.0
python-crontab==2.5.1
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.26.0
simplejson==3.17.3
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
